+++ NINA.Plugin.TargetScheduler/Controls/Reporting/ProfileSummaryView.xaml.cs
@@ -1,0 +1,11 @@
+using System.Windows.Controls;
+
+namespace NINA.Plugin.TargetScheduler.Controls.Reporting {
+
+    public partial class ProfileSummaryView : UserControl {
+
+        public ProfileSummaryView() {
+            InitializeComponent();
+        }
+    }
+}